add the re-add food diary feature

+++ app/src/main/java/org/cisnux/mydietary/data/remotes/responses/ServingResponse.kt
@@ -9,5 +9,5 @@ data class ServingResponse(
     val protein: Float,
     val carbohydrate: Float,
     val sugar: Float?,
-    val metricServingAmount: Int
+    val metricServingAmount: Float
 )
